+++ taxsystem/api/taxsystem/admin.py
@@ -1,9 +1,12 @@
+import logging
+
 from ninja import NinjaAPI
 
 from django.contrib.humanize.templatetags.humanize import intcomma
 from django.shortcuts import render
 
 from taxsystem.api.helpers import get_manage_corporation
+from taxsystem.api.taxsystem.helpers.administration import _delete_member
 from taxsystem.api.taxsystem.helpers.payments import _payments_actions
 from taxsystem.api.taxsystem.helpers.paymentsystem import (
     _get_has_paid_icon,
@@ -15,12 +18,11 @@ from taxsystem.api.taxsystem.helpers.statistics import (
     _get_statistics_dict,
 )
 from taxsystem.helpers import lazy
-from taxsystem.hooks import get_extension_logger
 from taxsystem.models.logs import AdminLogs
 from taxsystem.models.tax import Members, Payments, PaymentSystem
 from taxsystem.models.wallet import CorporationWalletDivision
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 class AdminApiEndpoints:
@@ -103,6 +105,13 @@ class AdminApiEndpoints:
             members = Members.objects.filter(corporation=corp)
 
             for member in members:
+                actions = _delete_member(
+                    corporation_id=corporation_id,
+                    member=member,
+                    perms=perms,
+                    request=request,
+                )
+
                 corporation_dict[member.character_id] = {
                     "character_id": member.character_id,
                     "character_portrait": lazy.get_character_portrait_url(
@@ -112,7 +121,7 @@ class AdminApiEndpoints:
                     "is_faulty": member.is_faulty,
                     "status": member.get_status_display(),
                     "joined": lazy.str_normalize_time(member.joined, hours=False),
-                    "actions": "",
+                    "actions": actions,
                 }
 
             output = []
@@ -134,19 +143,22 @@ class AdminApiEndpoints:
             if perms is False:
                 return 403, "Permission Denied"
 
-            payment_system = PaymentSystem.objects.filter(
-                corporation=corp
-            ).select_related("user", "user__profile", "user__profile__main_character")
+            payment_system = (
+                PaymentSystem.objects.filter(
+                    corporation=corp,
+                    user__profile__main_character__isnull=False,
+                )
+                .exclude(status=PaymentSystem.Status.MISSING)
+                .select_related(
+                    "user", "user__profile", "user__profile__main_character"
+                )
+            )
 
             payment_dict = {}
 
             for user in payment_system:
-                try:
-                    # Skip users without a main character
-                    character_id = user.user.profile.main_character.character_id
-                    character_name = user.user.profile.main_character.character_name
-                except AttributeError:
-                    continue
+                character_id = user.user.profile.main_character.character_id
+                character_name = user.user.profile.main_character.character_name
 
                 actions = _payment_system_actions(
                     corporation_id=corporation_id,