+++ taxsystem/app_settings.py
@@ -25,16 +25,20 @@ ZKILLBOARD_API_URL = "https://zkillboard.com/api/"
 ZKILLBOARD_BASE_URL_REGEX = r"^http[s]?:\/\/zkillboard\.com\/"
 ZKILLBOARD_KILLMAIL_URL_REGEX = r"^http[s]?:\/\/zkillboard\.com\/kill\/\d+\/"
 
-# Set Test Mode True or False
-
 # Set Naming on Auth Hook
 TAXSYSTEM_APP_NAME = clean_setting("TAXSYSTEM_APP_NAME", "Tax System")
 
 # If True you need to set up the Logger
 TAXSYSTEM_LOGGER_USE = clean_setting("TAXSYSTEM_LOGGER_USE", False)
 
-# Skip Dates for Audit
+# Task Settings
+# Global timeout for tasks in seconds to reduce task accumulation during outages.
+TAXSYSTEM_TASKS_TIME_LIMIT = clean_setting("TAXSYSTEM_TASKS_TIME_LIMIT", 7200)
 
+# Stale Time
+TAXSYSTEM_STALE_TIME = clean_setting("TAXSYSTEM_STALE_TIME", 3)
+
+# Skip Dates for Tasks
 # Member Skip Date in Hours
 TAXSYSTEM_CORP_MEMBERS_SKIP_DATE = clean_setting("TAXSYSTEM_CORP_MEMBERS_SKIP_DATE", 1)
 # Wallet Skip Date in Hours
@@ -46,4 +50,8 @@ TAXSYSTEM_CORP_PAYMENTS_SKIP_DATE = clean_setting(
 # Filter Skip Date in Hours
 TAXSYSTEM_CORP_PAYMENT_SYSTEM_SKIP_DATE = clean_setting(
     "TAXSYSTEM_CORP_PAYMENT_SYSTEM_SKIP_DATE", 1
+)
+# Payment Payday Skip Date in Hours
+TAXSYSTEM_CORP_PAYMENT_PAYDAY_SKIP_DATE = clean_setting(
+    "TAXSYSTEM_CORP_PAYMENT_PAYDAY_SKIP_DATE", 24
 )