+++ taxsystem/managers/logs_manager.py
@@ -1,10 +1,9 @@
 # Django
+import logging
+
 from django.db import models
 
-# AA Voices of War
-from taxsystem.hooks import get_extension_logger
-
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 class LogsQuerySet(models.QuerySet):