+++ taxsystem/models/tax.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 # Alliance Auth
-from allianceauth.authentication.models import User
+from allianceauth.authentication.models import OwnershipRecord, User
 from allianceauth.eveonline.models import EveAllianceInfo, EveCorporationInfo
 
 from taxsystem.managers.payment_manager import PaymentsManager, PaymentSystemManager
@@ -44,6 +44,8 @@ class OwnerAudit(models.Model):
     last_update_payments = models.DateTimeField(null=True, blank=True)
 
     last_update_payment_system = models.DateTimeField(null=True, blank=True)
+
+    last_update_payday = models.DateTimeField(null=True, blank=True)
 
     tax_amount = models.DecimalField(
         max_digits=16,
@@ -150,6 +152,7 @@ class PaymentSystem(models.Model):
         ACTIVE = "active", _("Active")
         INACTIVE = "inactive", _("Inactive")
         DEACTIVATED = "deactivated", _("Deactivated")
+        MISSING = "missing", _("Missing")
 
     name = models.CharField(
         max_length=100,
@@ -214,6 +217,10 @@ class PaymentSystem(models.Model):
     @property
     def is_deactivated(self) -> bool:
         return self.status == self.Status.DEACTIVATED
+
+    @property
+    def is_missing(self) -> bool:
+        return self.status == self.Status.MISSING
 
     @property
     def has_paid(self) -> bool:
@@ -290,6 +297,16 @@ class Payments(models.Model):
     @property
     def is_rejected(self) -> bool:
         return self.request_status == self.RequestStatus.REJECTED
+
+    @property
+    def character_id(self) -> int:
+        """Return the character ID of the user who made the payment or first OwnershipRecord."""
+        try:
+            character_id = self.account.user.profile.main_character.character_id
+        except AttributeError:
+            character = OwnershipRecord.objects.filter(user=self.account.user).first()
+            character_id = character.character.character_id
+        return character_id
 
     def __str__(self):
         return (