+++ taxsystem/task_helpers/etag_helpers.py
@@ -2,15 +2,15 @@
 Etag Helpers
 """
 
-import time
+import logging
 
-from bravado.exception import HTTPNotModified
+from bravado.exception import HTTPGatewayTimeout, HTTPNotModified
 
 from django.core.cache import cache
 
-from taxsystem.hooks import get_extension_logger
+from taxsystem.decorators import log_timing
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 MAX_ETAG_LIFE = 60 * 60 * 24 * 7  # 7 Days
 
@@ -19,9 +19,13 @@ class NotModifiedError(Exception):
     pass
 
 
+class HTTPGatewayTimeoutError(Exception):
+    pass
+
+
 def get_etag_key(operation):
     """Get ETag Key"""
-    return "etag-" + operation._cache_key()
+    return "taxsystem-" + operation._cache_key()
 
 
 def get_etag_header(operation):
@@ -86,16 +90,16 @@ def stringify_params(operation):
 
 
 def handle_etag_headers(operation, headers, force_refresh, etags_incomplete):
-    """Handle ETag headers"""
     if (
         get_etag_header(operation) == headers.headers.get("ETag")
         and not force_refresh
         and not etags_incomplete
     ):
-        logger.debug("Etag: No modified Data for %s", operation.operation.operation_id)
+        # Etag Match Cache Check
         raise NotModifiedError()
 
     if force_refresh:
+        # Remove Etag if force_refresh
         logger.debug(
             "ETag: Removing Etag %s F:%s - %s",
             operation.operation.operation_id,
@@ -104,6 +108,7 @@ def handle_etag_headers(operation, headers, force_refresh, etags_incomplete):
         )
         del_etag_header(operation)
     else:
+        # Save Etag
         logger.debug(
             "ETag: Saving Etag %s F:%s - %s",
             operation.operation.operation_id,
@@ -116,7 +121,6 @@ def handle_etag_headers(operation, headers, force_refresh, etags_incomplete):
 def handle_page_results(
     operation, current_page, total_pages, etags_incomplete, force_refresh
 ):
-    """Handle multiple page results and use Cache if possible"""
     results = []
     while current_page <= total_pages:
         operation.future.request.params["page"] = current_page
@@ -129,14 +133,36 @@ def handle_page_results(
             result, headers = operation.result()
             total_pages = int(headers.headers["X-Pages"])
             handle_etag_headers(operation, headers, force_refresh, etags_incomplete)
+
+            # Store results
             results += result
             current_page += 1
 
+            if not etags_incomplete and not force_refresh:
+                logger.debug(
+                    "ETag: No Etag %s - %s",
+                    operation.operation.operation_id,
+                    stringify_params(operation),
+                )
+                current_page = 1
+                results = []
+                etags_incomplete = True
+
         except (HTTPNotModified, NotModifiedError) as e:
-            logger.debug(e)
             if isinstance(e, NotModifiedError):
+                logger.debug(
+                    "ETag: Match Cache - Etag:%s, %s",
+                    operation.operation.operation_id,
+                    stringify_params(operation),
+                )
                 total_pages = int(headers.headers["X-Pages"])
             else:
+                logger.debug(
+                    "ETag: Match ESI - Etag: %s - %s ETag-Incomplete: %s",
+                    operation.operation.operation_id,
+                    stringify_params(operation),
+                    etags_incomplete,
+                )
                 total_pages = int(e.response.headers["X-Pages"])
 
             if not etags_incomplete:
@@ -145,12 +171,16 @@ def handle_page_results(
                 current_page = 1
                 results = []
                 etags_incomplete = False
+
+    if not etags_incomplete and not force_refresh:
+        raise NotModifiedError()
+
     return results, current_page, total_pages
 
 
+@log_timing(logger)
 def etag_results(operation, token, force_refresh=False):
     """Handle ETag results"""
-    _start_tm = time.perf_counter()
     operation.request_config.also_return_response = True
     if token:
         operation.future.request.headers["Authorization"] = (
@@ -160,6 +190,7 @@ def etag_results(operation, token, force_refresh=False):
         current_page = 1
         total_pages = 1
         etags_incomplete = False
+
         results, current_page, total_pages = handle_page_results(
             operation, current_page, total_pages, etags_incomplete, force_refresh
         )
@@ -172,11 +203,8 @@ def etag_results(operation, token, force_refresh=False):
             logger.debug("ETag: Not Modified %s", operation.operation.operation_id)
             set_etag_header(operation, e.response)
             raise NotModifiedError() from e
+        except HTTPGatewayTimeout as e:
+            logger.debug("ETag: Gateway Timeout %s", operation.operation.operation_id)
+            raise HTTPGatewayTimeoutError() from e
         handle_etag_headers(operation, headers, force_refresh, etags_incomplete=False)
-    logger.debug(
-        "ESI_TIME: OVERALL %s %s %s",
-        time.perf_counter() - _start_tm,
-        operation.operation.operation_id,
-        stringify_params(operation),
-    )
     return results