+++ taxsystem/task_helpers/general_helpers.py
@@ -2,21 +2,16 @@
 Core Helpers
 """
 
-from functools import wraps
-
-# pylint: disable=no-name-in-module
-from celery import signature
-from celery_once import AlreadyQueued
+import logging
 
 from esi.errors import TokenError
 from esi.models import Token
 
 from allianceauth.eveonline.models import EveCharacter
 
-from taxsystem.hooks import get_extension_logger
 from taxsystem.providers import esi
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 def get_token(character_id: int, scopes: list) -> Token:
@@ -87,43 +82,3 @@ def get_corp_token(corp_id, scopes, req_roles):
                 e,
             )
     return False
-
-
-def enqueue_next_task(chain):
-    """
-    Queue next task, and attach the rest of the chain to it.
-    """
-    while len(chain):
-        _t = chain.pop(0)
-        _t = signature(_t)
-        _t.kwargs.update({"chain": chain})
-        try:
-            _t.apply_async(priority=9)
-        except AlreadyQueued:
-            # skip this task as it is already in the queue
-            logger.debug("Skipping task as its already queued %s", _t)
-            continue
-        break
-
-
-def no_fail_chain(func):
-    """
-    Decorator to chain tasks provided in the chain kwargs regardless of task failures.
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        excp = None
-        _ret = None
-        try:
-            _ret = func(*args, **kwargs)
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            excp = e
-        finally:
-            _chn = kwargs.get("chain", [])
-            enqueue_next_task(_chn)
-            if excp:
-                raise excp
-        return _ret
-
-    return wrapper