+++ taxsystem/task_helpers/wallet_helpers.py
@@ -1,10 +1,11 @@
 """Wallet Helpers"""
 
+import logging
+
 from django.utils import timezone
 from eveuniverse.models import EveEntity
 
 from taxsystem.errors import DatabaseError
-from taxsystem.hooks import get_extension_logger
 from taxsystem.models.tax import (
     OwnerAudit,
 )
@@ -13,10 +14,14 @@ from taxsystem.models.wallet import (
     CorporationWalletJournalEntry,
 )
 from taxsystem.providers import esi
-from taxsystem.task_helpers.etag_helpers import NotModifiedError, etag_results
+from taxsystem.task_helpers.etag_helpers import (
+    HTTPGatewayTimeoutError,
+    NotModifiedError,
+    etag_results,
+)
 from taxsystem.task_helpers.general_helpers import get_corp_token
 
-logger = get_extension_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 def update_corporation_wallet_division(corp_id, force_refresh=False):
@@ -74,16 +79,21 @@ def update_corporation_wallet_division(corp_id, force_refresh=False):
                 update_corp_wallet_journal(
                     corp_id, division.get("division"), force_refresh=force_refresh
                 )  # inline not async
-
     except NotModifiedError:
         logger.debug(
             "No New wallet data for: %s",
             audit_corp.corporation.corporation_name,
         )
+    except HTTPGatewayTimeoutError:
+        # TODO Add retry method?
+        logger.debug(
+            "ESI Timeout skipping wallet for: %s",
+            audit_corp.corporation.corporation_name,
+        )
+        return ("ESI Timeout for %s:", audit_corp.corporation.corporation_name)
 
     audit_corp.last_update_wallet = timezone.now()
     audit_corp.save()
-
     return ("Finished wallet divs for: %s", audit_corp.corporation.corporation_name)
 
 
@@ -98,7 +108,7 @@ def update_corp_wallet_journal(corp_id, wallet_division, force_refresh=False):
     logger.debug(
         "Updating wallet transactions for: %s (Div: %s)",
         audit_corp.corporation.corporation_name,
-        division,
+        division.name,
     )
 
     req_scopes = [
@@ -176,7 +186,7 @@ def update_corp_wallet_journal(corp_id, wallet_division, force_refresh=False):
             logger.debug(
                 "Corp %s Div %s, Page: %s, New Transactions! %s, New Names %s",
                 corp_id,
-                wallet_division,
+                division.name,
                 current_page,
                 len(items),
                 _new_names,
@@ -195,10 +205,36 @@ def update_corp_wallet_journal(corp_id, wallet_division, force_refresh=False):
                 raise DatabaseError("DB Fail")
 
             current_page += 1
+        logger.debug(
+            "Finished wallet journal for: Div: %s Corp: %s",
+            division.name,
+            audit_corp.corporation.corporation_name,
+        )
+        return (
+            "Finished wallet journal for: Div: %s Corp: %s",
+            division.name,
+            audit_corp.corporation.corporation_name,
+        )
     except NotModifiedError:
         logger.debug(
             "No New wallet data for: Div: %s Corp: %s",
+            division.name,
             audit_corp.corporation.corporation_name,
-            wallet_division,
         )
-    return ("Finished wallet journal for: %s", audit_corp.corporation.corporation_name)
+        return (
+            "No New wallet data for: Div: %s Corp: %s",
+            division.name,
+            audit_corp.corporation.corporation_name,
+        )
+    except HTTPGatewayTimeoutError:
+        # TODO Add retry method?
+        logger.debug(
+            "ESI Timeout for Div: %s Corp: %s",
+            division.name,
+            audit_corp.corporation.corporation_name,
+        )
+        return (
+            "ESI Timeout for Div: %s Corp: %s",
+            division.name,
+            audit_corp.corporation.corporation_name,
+        )